% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHH.dowg.R
\name{gearPlotHH.dowg}
\alias{gearPlotHH.dowg}
\title{gearPlotHH.dowg Wing Spread vs. DoorSpread}
\usage{
gearPlotHH.dowg(
  Survey,
  years,
  quarter,
  c.int = 0.9,
  c.inta = 0.8,
  c.intb = 0.8,
  int.type = c("prediction", "confidence", "both"),
  es = FALSE,
  col1 = "darkblue",
  col2 = "steelblue2",
  getICES = T,
  pF = T,
  ti = T
)
}
\arguments{
\item{es}{if TRUE all legends and axis labels are in Spanish, if F in English}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{datHH:}{an HH data object with Survey, Year and Quarter columns, overrides Survey, Years, Quarters}

\item{c.int:}{the level (e.g. .95) of the band plotted when there is only one sweep length}

\item{c.inta:}{the level of the band plotted for the short sweeps in case there are two}

\item{c.intb:}{the level of the band plotted for the long set of sweeps}

\item{int.type:}{"prediction" (default) plots the range where an individual haul is expected to fall, does not shrink with more data and is the appropriate choice for flagging hauls with an abnormal gear geometry; "confidence" plots the range for the mean curve, which shrinks as data accumulates; "both" plots both bands}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if F title will not be included automatically and can be addedd later}
}
\value{
Produces a graph with WingSpread vs. DoorSpread, it also includes information on the ship, the time series used, the models and parameters estimated.
}
\description{
Plots Door Spread vs. Wing Spread behaviour and produces a model using lm. If there are DoorSpread and WingSpread values.
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
\dontrun{
gearPlotHH.dowg("SCOWCGFS",c(2014:2016),1,col1="darkblue",col2="steelblue3")
gearPlotHH.dowg("SCOWCGFS",c(2013:2016),4)
gearPlotHH.dowg("SCOROC",c(2013:2016),3)
gearPlotHH.dowg("NIGFS",c(2015:2016),1)
gearPlotHH.dowg("NIGFS",c(2006:2007,2009:2016),4)
gearPlotHH.dowg("IE-IGFS",c(2011:2016),4)
gearPlotHH.dowg("SP-PORC",c(2003:2015),3,c.int=.2)
gearPlotHH.dowg("FR-CGFS",c(2016:2018),4)
gearPlotHH.dowg("EVHOE",c(1997:2016),4)
gearPlotHH.dowg("SP-NORTH",c(2014:2016),4,col1="darkblue")
gearPlotHH.dowg("SP-ARSA",c(2014:2016),1,col1="darkblue",col2="steelblue2")
gearPlotHH.dowg("SP-ARSA",c(2014:2016),4)
gearPlotHH.dowg(damb,c(2014:2016),4,getICES=F,pF=F)
}
}
