% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHH.nodp.R
\name{gearPlotHH.nodp}
\alias{gearPlotHH.nodp}
\title{Function gearPlotHH.nodp to plot net opening vs. depth}
\usage{
gearPlotHH.nodp(
  Survey,
  years,
  quarter,
  c.inta = 0.8,
  c.intb = 0.3,
  int.type = c("prediction", "confidence", "both"),
  es = FALSE,
  col1 = "darkblue",
  col2 = "steelblue2",
  getICES = TRUE,
  pF = TRUE,
  ti = TRUE
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{c.inta:}{the level (e.g. .95) of the band plotted for all data if only one sweep length, and for the short sweeps in case there are two}

\item{c.intb:}{the level of the band plotted for the long set of sweeps}

\item{int.type:}{"prediction" (default) plots the range where an individual haul is expected to fall, does not shrink with more data and is the appropriate choice for flagging hauls with an abnormal gear geometry; "confidence" plots the range for the mean curve, which shrinks as data accumulates; "both" plots both bands}

\item{es:}{If T titles and labels are in Spanish, if FALSE in English}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if F title will not be included automatically and can be addedd later}
}
\value{
Produces Net Vertical opening vs. Depth plot it also includes information on the ship, the time series used (bottom fourth graph), the models and parameters estimated. The band level and int.type actually used are labelled on the plot: a single label when c.inta and c.intb match, or one label per curve, in that curve's color, when they differ.
}
\description{
Produces Net Vertical opening vs. Depth plot and a model with nls R function. Data are taken directly from DATRAS getting all the data from DATRAS using function getDATRAS from library(icesDatras)
it only produces plots for surveys with HH files uploaded in DATRAS
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
\dontrun{
gearPlotHH.nodp("SWC-IBTS",c(2014:2016),1,.07,.5,col1="darkblue",col2="steelblue2")
gearPlotHH.nodp("SWC-IBTS",c(2014:2016),1,.07,.5,col1="darkblue",col2="steelblue2",pF=F)
gearPlotHH.nodp("SWC-IBTS",c(2013:2016),4)
gearPlotHH.nodp("ROCKALL",c(2013:2016),3)
gearPlotHH.nodp("NIGFS",c(2005:2016),1)
gearPlotHH.nodp("NIGFS",c(2006:2007,2009:2016),4)
gearPlotHH.nodp("IE-IGFS",c(2011:2016),4,.8)
gearPlotHH.nodp("SP-PORC",c(2010:2016),3)
}
}
