% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHN21.dodp.R
\name{gearPlotHHN21.dodp}
\alias{gearPlotHHN21.dodp}
\title{Function gearPlotHHN21.dodp Door Spread versus Depth N21 Spanish North with two vessels}
\usage{
gearPlotHHN21.dodp(
  Survey = "SP-NORTH",
  years = 2021,
  quarter = 4,
  c.int = 0.95,
  c.inta = 0.95,
  c.intb = 0.95,
  es = FALSE,
  col1 = "darkblue",
  col2 = "red",
  getICES = TRUE,
  pF = TRUE,
  ti = TRUE
)
}
\arguments{
\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be plotted}

\item{c.int:}{the confidence interval to be used in the confint function}

\item{c.inta:}{the confidence interval to be used in the confint function for all data if only one sweep length, and for the short sweeps in case there are two}

\item{c.intb:}{the confidence interval to be used in the confint function for the long set of sweeps.}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if F title will not be included automatically and can be addedd later}
}
\value{
Produces a graph DoorSpread vs. Depth, it also includes information on the ship, the time series used, the models and parameters estimated.
}
\description{
Produces a DoorSpread vs. Depth plot and model with nls R function. Data are taken directly from DATRAS using function getDATRAS from library(icesDatras)
it only produces plots for surveys with HH files uploaded in DATRAS
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
gearPlotHHN21.dodp("SP-NORTH",c(2021),3,.8,.3,col1="darkblue",col2="red")
}
