% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHN21.nodp.R
\name{gearPlotHHN21.nodp}
\alias{gearPlotHHN21.nodp}
\title{Function gearPlotHH.nodp to plot net opening vs. depth for the Spanish N21 Survey}
\usage{
gearPlotHHN21.nodp(
  Survey = "SP-NORTH",
  years = 2021,
  quarter = 4,
  c.inta = 0.8,
  c.intb = 0.3,
  es = FALSE,
  col1 = "darkblue",
  col2 = "red",
  incl2 = TRUE,
  getICES = TRUE,
  pF = TRUE,
  ti = TRUE
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{c.inta:}{the confidence interval to be used in the confint function for all data if only one sweep length, and for the short sweeps in case there are two}

\item{c.intb:}{the confidence interval to be used in the confint function for the long set of sweeps.}

\item{es:}{If TRUE all labels and legends are in Spanish, if FALSE in English}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if F title will not be included automatically and can be addedd later}
}
\value{
Produces Net Vertical opening vs. Depth plot it also includes information on the ship, the time series used (bottom fourth graph), the models and parameters estimated.
}
\description{
Produces Net Vertical opening vs. Depth plot and a model with nls R function. Data are taken directly from DATRAS getting all the data from DATRAS using function getDATRAS from library(icesDatras)
it only produces plots for surveys with HH files uploaded in DATRAS
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
gearPlotHH.nodp("SWC-IBTS",c(2014:2016),1,.07,.5,col1="darkblue",col2="steelblue2")
gearPlotHH.nodp("SWC-IBTS",c(2014:2016),1,.07,.5,col1="darkblue",col2="steelblue2",pF=F)
gearPlotHH.nodp("SWC-IBTS",c(2013:2016),4)
gearPlotHH.nodp("ROCKALL",c(2013:2016),3)
gearPlotHH.nodp("NIGFS",c(2005:2016),1)
gearPlotHH.nodp("NIGFS",c(2006:2007,2009:2016),4)
gearPlotHH.nodp("IE-IGFS",c(2011:2016),4,.8)
gearPlotHH.nodp("SP-PORC",c(2010:2016),3)
gearPlotHH.nodp("FR-CGFS",c(2014:2016),4)
gearPlotHH.nodp("EVHOE",c(1997:2015),4)
gearPlotHH.nodp("SP-NORTH",c(2014:2016),4,col1="darkblue",col2="yellow")
gearPlotHH.nodp("SP-ARSA",c(2014:2016),1)
gearPlotHH.nodp("SP-ARSA",c(2014:2016),4)
gearPlotHH.nodp(getICES=F,damb,c(2014:2016),4,pF=F)
gearPlotHH.nodp(getICES=F,Survey=damb,years=c(2014:2016),quarter=4,pF=F)
}
