% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHN21.wrpdp.R
\name{gearPlotHHN21.wrpdp}
\alias{gearPlotHHN21.wrpdp}
\title{Function gearPlotHH.wrpdp plots warp length vs. Depth behaviour}
\usage{
gearPlotHHN21.wrpdp(
  Survey = "SP-NORTH",
  years = 2021,
  quarter = 4,
  incl2 = TRUE,
  line = TRUE,
  c.inta = 0.95,
  int.type = c("prediction", "confidence", "both"),
  es = FALSE,
  col1 = "darkblue",
  col2 = "red",
  getICES = TRUE,
  pF = TRUE,
  ti = TRUE
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{line:}{includes a regression line between Warp and depth and the formula of the linear regression. If F the line is omited}

\item{c.inta:}{the level (e.g. .95) of the band plotted, passed to \code{\link{gearBand}}}

\item{int.type:}{"prediction" (default) plots the range where an individual haul is expected to fall, does not shrink with more data and is the appropriate choice for flagging hauls with an abnormal gear geometry; "confidence" plots the range for the mean curve, which shrinks as data accumulates; "both" plots both bands}

\item{es:}{if TRUE labels and axes labels, titles legends in Spanish, if FALSE in English}

\item{col1:}{the color of the points, last year fill and previous years empty symbol}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if T includes autmoatically the title, F leaves it blank an can be added later.}
}
\value{
Produces a graph Warp length vs. Depth for the years selected.
}
\description{
Produces a warplength vs. DoorSpread plot and a model with lm unction. Data are taken directly from DATRAS using function getDATRAS from library(icesDatras)
It only produces plots for surveys with HH files uploaded in DATRAS
Since sweeps length does not affect the warp that is decided by the chief scientist, there are no differences depending on sweeps
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
\dontrun{
gearPlotHH.wrpdp("SWC-IBTS",c(2011:2016),1,c.inta=.95,col1="darkblue",pF=F)
gearPlotHH.wrpdp("SWC-IBTS",c(2013:2016),4)
gearPlotHH.wrpdp(damb,c(2014:2016),4,getICES=F,pF=F)
}
}
