% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHNS.dodp.R
\name{gearPlotHHNS.dodp}
\alias{gearPlotHHNS.dodp}
\title{Function gearPlotHHNS.dodp Door Spread versus Depth for a concrete country within the NS-IBTS survey}
\usage{
gearPlotHHNS.dodp(
  Survey = "NS-IBTS",
  years,
  quarter,
  country,
  c.inta = 0.8,
  c.intb = 0.3,
  int.type = c("prediction", "confidence", "both"),
  col1 = "darkblue",
  col2 = "steelblue2",
  getICES = T,
  pF = T
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be plotted}

\item{country:}{The country chosen to be plotted (checks if it's available in the HH file)}

\item{c.inta:}{the level (e.g. .95) of the band plotted for all data if only one sweep length, and for the long sweeps in case there are two}

\item{c.intb:}{the level of the band plotted for the short set of sweeps in case there are two}

\item{int.type:}{"prediction" (default) plots the range where an individual haul is expected to fall, does not shrink with more data and is the appropriate choice for flagging hauls with an abnormal gear geometry; "confidence" plots the range for the mean curve, which shrinks as data accumulates; "both" plots both bands}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}
}
\value{
Produces a graph DoorSpread vs. Depth, it also includes information on the ship, the time series used, the models and parameters estimated.
}
\description{
Produces a DoorSpread vs. Depth plot and model with nls R function. Data are taken directly from DATRAS using function getHHdata from library(icesDatras)
it only produces plots for surveys with HH files uploaded in DATRAS
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
\dontrun{
gearPlotHHNS.dodp("NS-IBTS",c(2014:2017),3,"SCO",.8,.3,col1="darkblue",col2="darkgreen")
}
}
