% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHNS.wgdo.R
\name{gearPlotHHNS.wgdo}
\alias{gearPlotHHNS.wgdo}
\title{gearPlotHH.wgdo Door Spread vs. WingSpread}
\usage{
gearPlotHHNS.wgdo(
  Survey = "NS-IBTS",
  years,
  quarter,
  country,
  col1 = "darkblue",
  col2 = "steelblue2",
  getICES = T,
  pF = T
)
}
\arguments{
\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{country:}{The country chosen to be plotted (checks if it's available in the HH file)}

\item{datHH:}{an HH data object with Survey, Year and Quarter columns, overrides Survey, Years, Quarters}

\item{col1:}{color for the symbols and lines for the whole set if only one set of sweeps are used, and for the data from the long set of sweeps.}

\item{col2:}{color for the symbols and lines for the data from the short sweeps in case there are two.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}

\item{ti:}{if T includes autmoatically the title, F leaves it blank an can be added later.}
}
\value{
Produces a graph with DoorSpread vs. WingSpread, it also includes information on the ship, the time series used, the models and parameters estimated.
}
\description{
Plots Door Spread vs. Wing Spread behaviour and produces a model using lm. If there are DoorSpread and WingSpread values.
If there are two different sweeps in the data, produces a model for each sweep length.
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
gearPlotHHNS.wgdo("NS-IBTS",c(2016:2017),1,"SWE")
gearPlotHHNS.wgdo("NS-IBTS",c(2015:2017),1,"SWE")
}
