% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatras2.R
\name{getDatras2}
\alias{getDatras2}
\title{Function getDatras2 modification of function getDATRAS to give the possibility of not getting the list of files downloaded}
\usage{
getDatras2(record = "HH", survey, years, quarters, listData = FALSE)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}
}
\value{
Downloads from DATRAS the HH files from the survey and the years selected.
}
\description{
Function getDatras2 modification of function getDATRAS to give the possibility of not getting the list of files downloaded
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
getDatras2("HH","SP-NORTH",c(2014:2016),4,listData=FALSE)
}
