% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTSNeAtl_map.R
\name{IBTSNeAtl_map}
\alias{IBTSNeAtl_map}
\title{Function IBTSNeAtl_map plots the map with all the surveys \strong{STILL ON DEVELOPMENT}}
\usage{
IBTSNeAtl_map(
  nl = 60.5,
  sl = 36,
  xlims = c(-18, 3),
  leg = TRUE,
  legpos = c("bottomright"),
  cex.leg = 0.9,
  dens = 300,
  load = TRUE,
  ICESdiv = TRUE,
  ICESrect = FALSE,
  ICESlab = FALSE,
  ICESlabcex = 0.8,
  NS = FALSE,
  bathy = TRUE,
  bw = FALSE,
  axlab = 0.8,
  bords = TRUE,
  lwdl = 0.1,
  shpdir = "c:/GitHubRs/shapes/",
  places = FALSE,
  minpop = 2e+05,
  graf = FALSE,
  xpng = 1200,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{nl}{= 61.5 northernmost limit of the map}

\item{sl}{= 35 Southernmost limit of the map}

\item{leg}{= TRUE if TRUE includes the legend with the colors of the surveys}

\item{legpos}{defines where the legend is placed}

\item{cex.leg}{= .7 Size of the legend}

\item{dens}{= 30 density of the shading lines for all the surveys}

\item{load}{= T or F to load all the shapes files.}

\item{ICESdiv}{= TRUE if TRUE plots the IBTS divisions behind the shapefiles}

\item{ICESrect}{= FALSE if TRUE plots the lines of the ICES statistic rectangles}

\item{ICESlab}{= FALSE if TRUE plots labs por ICES rectangles}

\item{ICESlabcex}{tamano del ICESlab en cex, .5 por defecto subirlo si se quiere mas grande}

\item{NS}{= FALSE if TRUE includes the ICES rectangles only for the North Sea area}

\item{bathy}{= TRUE if TRUE plots the isobaths under the behind the shapefiles}

\item{bw}{= True plots the map with land in grey, if F in light brown (burlywood3)}

\item{axlab}{=1 size of axis labs}

\item{bords}{= TRUE plots the borders o the countries, FALSE leaves dashed lines}

\item{shpdir}{= path to the folder with the shapefiles}

\item{places}{= if T adds towns in the maps}

\item{minpop}{= adds the limit of population of cities (places) to plot.}

\item{graf}{si F el grafico sale en la pantalla, si nombre fichero va a fichero en el directorio de trabajo del Rstudio ver getwd()}

\item{xpng}{width archivo png si graf es el nombre del fichero}

\item{ypng}{height archivo png si graf es el nombre del fichero}

\item{ppng}{points png archivo si graf es el nombre del fichero}

\item{axlab=}{.8 decides the size of the axis numbers}
}
\description{
#' @description
to Cadiz, Still needs to include the shapefiles within the package, or right references.
}
\details{
The default limits for the IBTS including the North Sea or not, but the script could
be used to get maps all arround the world playing with nl,sl and xlims
}
\examples{
IBTSNeAtl_map(dens=0,nl=45,leg=F,load=TRUE,ICESrect = T,graf="MapIBTS");text(stat_y~stat_x,Area,labels=ICESNAME,cex=.8,font=4);text(stat_y~stat_x,Area,labels=Area,cex=.6,pos=1,col=2)
}
