% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTSNeAtl_map_Porc.R
\name{IBTSNeAtl_map_porc}
\alias{IBTSNeAtl_map_porc}
\title{Function IBTSNeAtl_map_porc plots the map with only Porcupine}
\usage{
IBTSNeAtl_map_porc(
  nl = 54.5,
  sl = 50.5,
  xlims = c(-15.5, -8.5),
  leg = FALSE,
  es = FALSE,
  places = T,
  cex.leg = 0.7,
  dens = 30,
  ICESdiv = TRUE,
  ICESrect = FALSE,
  bathy = TRUE,
  out = "def",
  nfile = "NeAtlIBTS_map_porc",
  lwdl = 0.1,
  load = TRUE,
  shpdir = "c:/GitHubRs/shapes/"
)
}
\arguments{
\item{nl}{= 54.5 northernmost limit of the map}

\item{sl}{= 50.5 Southernmost limit of the map}

\item{leg}{= TRUE if TRUE includes the legend with the colors of the surveys}

\item{es}{= F if F texts in English, if TRUE in Spanish}

\item{places}{=T if T incluye letreros de Irlanda y Galway}

\item{cex.leg}{= .7 Size of the legend}

\item{dens}{= 30 density of the shading lines for all the surveys}

\item{ICESdiv}{= TRUE if TRUE plots the IBTS divisions behind the shapefiles}

\item{ICESrect}{= FALSE if TRUE plots the lines of the ICES statistic rectangles}

\item{bathy}{= TRUE if TRUE plots the isobaths under the behind the shapefiles}

\item{out}{= format of the output, can be "def" default device, "pdf", "tiff" or "png"}

\item{nfile}{= name for the output file}

\item{load}{= T or F to load all the shapes files}

\item{shpdir}{= path to the folder with the shapefiles}
}
\description{
Produces a map from the shapefiles that define the IBTSNeAtl solo Porcupine
}
\examples{
IBTSNeAtl_map_porc(dens=0,leg=F,ICESrect = T);text(stat_y~stat_x,Area,labels=ICESNAME,cex=.8,font=4);text(stat_y~stat_x,Area,labels=Area,cex=.6,pos=1,col=2)
}
