% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTSurveys.R
\docType{data}
\name{IBTSurveys}
\alias{IBTSurveys}
\title{List of current IBTS surveys}
\format{
Vector with the codes of thes surveys
}
\usage{
IBTSsurv
}
\description{
List with the IBTS surveys not discontinued
}
\concept{Data info}
