% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfoHH.R
\name{InfoHH}
\alias{InfoHH}
\title{Function InfoHH gives the information on Sweep lengths for the years selected
Data are taken directly from DATRAS getting all the data from DATRAS using function getHHdata from library(icesDatras)}
\usage{
InfoHH(Survey, years, quarter, country = NA, getICES = T)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{country:}{The country chosen to be plotted (checks if it's available in the HH file)}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDATRAS package. If F it analyzes the HH dataframe just with the available years independently of the years and quarter asked for.}
}
\value{
Produces a graph with WingSpread vs. Depth. it also includes information on the ship, the time series used, the models and parameters estimated.
}
\description{
Function InfoHH gives the information on Sweep lengths for the years selected
Data are taken directly from DATRAS getting all the data from DATRAS using function getHHdata from library(icesDatras)
}
\details{
Surveys available in DATRAS: i.e. NS-IBTS, SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
InfoHH(Survey="SWC-IBTS",years=c(2011:2016),quarter=4)
InfoHH(damb,c(2014:2016),4,getICES=F)
}
