% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapLengths_nc.R
\name{MapLengths_nc}
\alias{MapLengths_nc}
\title{Function SplitLengths extracts from DATRAS the data to produce the IBTS maps with two length ranges for a fix set of species see file SpeciesCodes.csv}
\usage{
MapLengths_nc(
  esp,
  dtSurv,
  dtyear,
  dtq,
  tmin,
  tmax,
  add = FALSE,
  ti = TRUE,
  subti = TRUE,
  colo = "red",
  bw = TRUE,
  save.dat = FALSE,
  out.dat = FALSE,
  zeros = FALSE,
  onlyVal = F,
  cexleg = 1,
  escmult = 1,
  escCPUE = NA
)
}
\arguments{
\item{datSurvey:}{The Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{dtyear:}{year to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{dtq:}{the quarter of the survey to be downloaded}

\item{esp:}{species to be included in the resulting map if plot=True}

\item{adj:}{weight abundance within species to the maximum abundance of the specie in the survey}

\item{escmult:}{regulates point size}

\item{cexleg:}{size of the legends and the texts in the graph}

\item{plot:}{if TRUE a map with the data selected in the species esp is presented}

\item{colo:}{bg of the color of the points}

\item{ti:}{includes a title with the name of the species in the plot}

\item{bw:}{if T map in black and white}

\item{save.dat:}{by default set to FALSE, if TRUE saves the data in a file: IBTSdataSURVEYyrQX.csv}

\item{out.dat:}{by default set to FALSE, if TRUE gives the data as output of the function}

\item{zeros:}{if TRUE the map includes small points with the hauls with out catch of the species}
}
\value{
during the calculations shows the types of data in the file "C", "P", "R", And the species of the set present in the Survey/year
}
\description{
Function SplitLengths extracts from DATRAS the data to produce the IBTS maps with two length ranges for a fix set of species see file SpeciesCodes.csv
}
\details{
Surveys available in DATRAS recent years and not discontinued: EVHOE, FR-CGFS,
FR-WCGFS, IE-IAMS, IE-IGFS, NIGFS, NS-IBTS, PT-IBTS, SCOROC, SCOWCGFS, SP-ARSA, SP-NORTH, SP-PORC
}
\examples{
MapLengths("SP-NORTH",2024,4,esp="HKE",tmin=0,tmax=20,zeros=T,add=F)
MapLengths("NS-IBTS",2023,3,esp="MEG",tmin=20,tmax=50,zeros=T,add=F)
IBTSNeAtl_map(load=F,leg=F,xlims=c(-16,13),bw=T);MapLengths("NS-IBTS",2023,3,esp="HKE",tmin=0,tmax=21,zeros=T,add=T)
}
