% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcHaulsPosit.R
\name{qcHaulsPosit}
\alias{qcHaulsPosit}
\title{Function qcHaulsPosit plots warp length vs. Depth behaviour}
\usage{
qcHaulsPosit(
  Survey = "NS-IBTS",
  years,
  quarter,
  col1 = "red",
  ti = TRUE,
  Hpoints = FALSE,
  Nhauls = FALSE,
  getICES = TRUE,
  esc.mult = 1,
  graf = FALSE,
  xpng = 800,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{graf}{if FALSE the graph goes to screen, if its a file name (i.e. "graf") a .png file with that name is created and a message with location (wd) is shown in screen.}

\item{xpng}{width file png if graf is the name of the file}

\item{ypng}{height file png if graf is the name of the file}

\item{ppng}{points png parameter if graf is the name of the file}

\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{country:}{The country chosen to be plotted (checks if it's available in the HH file)}

\item{col1:}{the color of the lines}

\item{nhauls:}{if T shows the number of the hauls on the plot}

\item{ti=}{if T includes the title of the graph with country, survey, year and quarter}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}
}
\value{
Produces a map with segments of the hauls performed in that survey.
}
\description{
Produces a warplength vs. DoorSpread plot and a model with lm unction. Data are taken directly from DATRAS using function getHHdata from library(icesDatras)
It only produces plots for surveys with HH files uploaded in DATRAS
Since sweeps length does not affect the warp that is decided by the chief scientist, there are no differences depending on sweeps
Maps try to use ShootLat and ShootLong cause old data from NS have a lot of -9 (NA) values in the NS-IBTS
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
gearPlotdumb.wrpdp("NS-IBTS",c(2014:2017),1,"SWE")
}
