% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesInHL.R
\name{SpeciesInHL}
\alias{SpeciesInHL}
\title{Species list in DATRAS HL}
\usage{
SpeciesInHL(Survey, Year, Quarter)
}
\arguments{
\item{Survey}{survey name from those present in DATRAS}

\item{Year}{Year of the survey}

\item{Quarter}{Quarter of the survey}
}
\value{
A vector with the species list in the HL
}
\description{
Species list in DATRAS HL
}
\examples{
\dontrun{SpeciesInHL("SP-NORTH", 2025, 4)}
}
