% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitLengthsESP.R
\name{SplitLengthsESP}
\alias{SplitLengthsESP}
\title{Function SplitLengths extracts from DATRAS the data to produce the IBTS maps with two length ranges for a fix set of species see file SpeciesCodes.csv}
\usage{
SplitLengthsESP(
  datSurvey,
  dtyear,
  dtq,
  esp = "Merluccius merluccius",
  aphia = FALSE,
  L_Split = 21,
  plot = TRUE,
  legpos = "bottomright",
  ti = TRUE,
  save.dat = FALSE,
  out.dat = FALSE,
  zeros = FALSE
)
}
\arguments{
\item{datSurvey:}{The Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{dtyear:}{year to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{dtq:}{the quarter of the survey to be downloaded}

\item{esp:}{species to be included in the resulting map if plot=True, name as scientific upto date with WoRMS}

\item{plot:}{if TRUE a map with the data selected in the species esp is presented}

\item{legpos:}{by default "bottomright", but could be should be one of "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right", "center"}

\item{ti:}{includes a title with the name of the species in the plot}

\item{save.dat:}{by default set to FALSE, if TRUE saves the data in a file: IBTSdataSURVEYyrQX.csv}

\item{out.dat:}{by default set to FALSE, if TRUE gives the data as output of the function}

\item{zeros:}{if TRUE the map includes small points with the hauls with out catch of the species}
}
\value{
during the calculations shows the types of data in the file "C", "P", "R", And the species of the set present in the Survey/year
}
\description{
Function SplitLengths extracts from DATRAS the data to produce the IBTS maps with two length ranges for a fix set of species see file SpeciesCodes.csv
}
\details{
Surveys available in DATRAS recent years and not discontinued: EVHOE, FR-CGFS,
FR-WCGFS, IE-IAMS, IE-IGFS, NIGFS, NS-IBTS, PT-IBTS, SCOROC, SCOWCGFS, SP-ARSA, SP-NORTH, SP-PORC
}
\examples{
\dontrun{
SplitLengths("NS-IBTS",2023,3,esp="HKE",zeros=T)
SplitLengths("NS-IBTS",2023,3,esp="MEG",zeros=T)
}
}
