% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearPlotHHNS.wrpdp.R
\name{gearPlotHHNS.wrpdp}
\alias{gearPlotHHNS.wrpdp}
\title{Function gearPlotHHNS.wrpdp plots warp length vs. Depth behaviour}
\usage{
gearPlotHHNS.wrpdp(
  Survey = "NS-IBTS",
  years,
  quarter,
  country,
  line = T,
  c.inta = 0.85,
  col1 = "darkblue",
  getICES = T,
  pF = T
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{country:}{The country chosen to be plotted (checks if it's available in the HH file)}

\item{line:}{includes a regression line between Warp and depth and the formula of the linear regression. If F the line is omited}

\item{c.inta:}{the confidence interval to be used in the predict.lm function}

\item{col1:}{the color of the points, last year fill and previous years empty symbol}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}

\item{pF:}{takes out the points and leaves only the lines in the graphs}
}
\value{
Produces a graph Warp length vs. Depth for the years selected.
}
\description{
Produces a warplength vs. DoorSpread plot and a model with lm unction. Data are taken directly from DATRAS using function getHHdata from library(icesDatras)
It only produces plots for surveys with HH files uploaded in DATRAS
Since sweeps length does not affect the warp that is decided by the chief scientist, there are no differences depending on sweeps
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
gearPlotHHNS.wrpdp("NS-IBTS",c(2014:2017),1,"SWE")
}
