% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcHaulsDist.R
\name{qcHaulsDist}
\alias{qcHaulsDist}
\title{Function qcHaulsDist checks consistency between hauls distance, speed and course parameters}
\usage{
qcHaulsDist(
  Survey = "NS-IBTS",
  years,
  quarter,
  pc.error = 2,
  error.rb = TRUE,
  allHauls = FALSE,
  plots = TRUE,
  getICES = TRUE,
  error = c("Dist"),
  esc.mult = 1,
  graf = FALSE,
  xpng = 800,
  ypng = 800,
  ppng = 15
)
}
\arguments{
\item{pc.error}{Acceptable error rate for not displaying hauls as erroneous}

\item{error.rb}{if FALSE does not take into account course errors}

\item{allHauls}{Default is FALSE if TRUE lists values for all hauls, if not only those where pc.error>error}

\item{graf}{if FALSE the graph goes to screen, if its a file name (i.e. "graf") a .png file with that name is created and a message with location (wd) is shown in screen.}

\item{xpng}{width file png if graf is the name of the file}

\item{ypng}{height file png if graf is the name of the file}

\item{ppng}{points png parameter if graf is the name of the file}

\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{getICES:}{Should the data be downloaded from DATRAS? If T, default, the data are taken from DATRAS through the icesDatras package.}
}
\value{
data.frame with survey, haul, distance and estimated values different errors and hours of dawn/sunrise/dusk
}
\description{
Produces different plots comparing expected values and those included.
Data are taken directly from DATRAS using function getDATRAS from library(icesDatras)
It only produces plots for surveys with HH files uploaded in DATRAS
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
qcHaulsDist("SP-NORTH",c(2024),4,error="Speed"); qcHaulsDist("IE-IAMS",2024,quarter=c(1,1),error="Course")
}
\references{
distHaversine function gives the haversine calculation of distance between two geographic points \code{\link[geosphere]{distHaversine}}
}
\seealso{
Other quality control: 
\code{\link{safe_getDATRAS}()}
}
\concept{quality control}
