% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_getDATRAS.R
\name{safe_getDATRAS}
\alias{safe_getDATRAS}
\title{Function safe_getDATRAS gets data from DATRAS and shows original messages to the users but inside R and shiny}
\usage{
safe_getDATRAS(
  datatype = "HH",
  survey,
  years,
  quarter,
  context = c("console", "shiny")
)
}
\arguments{
\item{Survey:}{either the Survey to be downloaded from DATRAS (see details), or a data frame with the HH information with  the DATRAS HH format  and the years and quarter selected in years and quarter}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in steelblue2, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}

\item{context:}{evaluates if running within a normal script or within shiny app}
}
\value{
data.frame withtthe DATRAS data for the selected survey, year and quaerter
}
\description{
Downloads data from DATRAS or sends its errors
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
safe_getDATRAS("SP-NORTH",2022,4,context="console")
}
\seealso{
Other quality control: 
\code{\link{qcHaulsDist}()}
}
\concept{quality control}
