% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesCAperYear.R
\name{SpeciesCAperYear}
\alias{SpeciesCAperYear}
\title{Function SpeciesCAperYear Takes a CA file for one survey and produces a table with number of biological samples per year and species}
\usage{
SpeciesCAperYear(Survey, years, quarter)
}
\arguments{
\item{Survey:}{Surveys available in DATRAS see details}

\item{years:}{years to be downloaded and used, had to be available in DATRAS. The time series will be ploted in grey dots, last year in yellow, it depends on the order of years, not the actual chronological year.}

\item{quarter:}{the quarter of the survey to be ploted}
}
\description{
Function SpeciesCAperYear Takes a CA file for one survey and produces a table with number of biological samples per year and species
}
\details{
Surveys available in DATRAS: i.e. SWC-IBTS, ROCKALL, NIGFS, IE-IGFS, SP-PORC, FR-CGFS, EVHOE, SP-NORTH, PT-IBTS and SP-ARSA
}
\examples{
SpeciesCAperYear("SP-NORTH",c(2014:2016),4)
}
